+++ src/main/java/com/kummit/api_server/enums/CodingTier.java
@@ -1,0 +1,3 @@
+package com.kummit.api_server.enums;
+
+public enum CodingTier { BRONZE, SILVER, GOLD, EXPERT }
