+++ src/main/java/com/kummit/api_server/enums/Status.java
@@ -1,0 +1,3 @@
+package com.kummit.api_server.enums;
+
+public enum Status { ACTIVE, DELETED, SUSPENDED }
